Improve naming of main class
